#459: Moved and renamed the file containing the PassiveTargetChanceChangesParser
 The Parser was not modified.
